Added filtering functionality to polygon_matcher class, moving get_plotting_polygons with it. Changed notebook examples to accomodate for this change.
